update benchmarks settings, split evaluation test
